Move UI code into ui folder
